+++ layers/CMSSchema/packages/custompost-tag-mutations/src/ObjectTypeResolverPickers/CustomPostDoesNotExistErrorPayloadObjectTypeResolverPicker.php
@@ -4,8 +4,9 @@ declare(strict_types=1);
 
 namespace PoPCMSSchema\CustomPostTagMutations\ObjectTypeResolverPickers;
 
-use PoPCMSSchema\CustomPostTagMutations\TypeResolvers\UnionType\RootSetTagsOnCustomPostMutationErrorPayloadUnionTypeResolver;
 use PoPCMSSchema\CustomPostMutations\ObjectTypeResolverPickers\AbstractCustomPostDoesNotExistErrorPayloadObjectTypeResolverPicker;
+use PoPCMSSchema\CustomPostTagMutations\TypeResolvers\UnionType\GenericCustomPostSetTagsMutationErrorPayloadUnionTypeResolver;
+use PoPCMSSchema\CustomPostTagMutations\TypeResolvers\UnionType\RootSetTagsOnCustomPostMutationErrorPayloadUnionTypeResolver;
 use PoP\ComponentModel\TypeResolvers\UnionType\UnionTypeResolverInterface;
 
 class CustomPostDoesNotExistErrorPayloadObjectTypeResolverPicker extends AbstractCustomPostDoesNotExistErrorPayloadObjectTypeResolverPicker
@@ -17,6 +18,7 @@ class CustomPostDoesNotExistErrorPayloadObjectTypeResolverPicker extends Abstrac
     {
         return [
             RootSetTagsOnCustomPostMutationErrorPayloadUnionTypeResolver::class,
+            GenericCustomPostSetTagsMutationErrorPayloadUnionTypeResolver::class,
         ];
     }
 }