+++ layers/Engine/packages/root-wp/src/Module.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace PoP\RootWP;
 
-use Brain\Cortex;
 use PoP\Root\App;
 use PoP\Root\Environment;
 use PoP\Root\Module as RootModule;
@@ -99,7 +98,9 @@ class Module extends AbstractModule
          *     "brain/cortex": "1.0.0-alpha.9"
          *   }
          *   ```
+         *
+         * @todo: Uncomment this line below if decide to use Cortex again
          */
-        Cortex::boot();
+        // Cortex::boot();
     }
 }