+++ layers/GatoGraphQLForWP/plugins/gatographql/src/Log/Controllers/Internal/Caching/CacheHelper.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace GatoGraphQL\GatoGraphQL\Log\Controllers\Internal\Caching;
+
+/**
+ * @phpcs:ignoreFile
+ */
+class CacheHelper
+{
+    use CacheNameSpaceTrait;
+}