+++ layers/GatoGraphQLForWP/plugins/gatographql/src/PluginEnvironment.php
@@ -72,9 +72,4 @@ class PluginEnvironment
     {
         return static::getGatoGraphQLDynamicFileStorageDir() . \DIRECTORY_SEPARATOR . 'logs';
     }
-
-    public static function getLogsFilePath(string $filename): string
-    {
-        return static::getLogsDir() . \DIRECTORY_SEPARATOR . $filename;
-    }
 }