+++ layers/GatoGraphQLForWP/plugins/testing-schema/src/Overrides/Log/SystemLogger.php
@@ -1,0 +1,24 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GatoGraphQL\TestingSchema\Overrides\Log;
+
+use PoPSchema\Logger\Log\SystemLogger as UpstreamSystemLogger;
+use GatoGraphQL\TestingSchema\Constants\CustomHeaders;
+
+class SystemLogger extends UpstreamSystemLogger
+{
+    use LoggerTrait;
+
+    /**
+     * Send the error to the response headers,
+     * so we can test it
+     */
+    public function log(string $message): void
+    {
+        parent::log($message);
+
+        $this->sendCustomHeader($message, CustomHeaders::GATOGRAPHQL_ERRORS);
+    }
+}