+++ layers/Schema/packages/logger/src/Constants/LoggerContext.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PoPSchema\Logger\Constants;
+
+class LoggerContext
+{
+    public const LOG_ENTRY_CONTEXT_SEPARATOR = 'CONTEXT: ';
+}