+++ layers/Schema/packages/logger/src/Constants/LoggerSigns.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PoPSchema\Logger\Constants;
+
+class LoggerSigns
+{
+    public final const ERROR = '🔴';
+    public final const WARNING = '🟡';
+    public final const INFO = '🔵';
+    public final const DEBUG = '🟢';
+}