+++ layers/Schema/packages/logger/src/Log/LoggerSources.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PoPSchema\Logger\Log;
+
+class LoggerSources
+{
+    public const INFO = 'info';
+}