+++ layers/GatoGraphQLForWP/plugins/gatographql/CHANGELOG.md
@@ -6,6 +6,28 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
 ## 13.0.0 - DATE
 
+### Breaking changes
+
+- Replaced methods in `Logger`: `logInfo` and `logError` => `log(severity: ...)` (#3093)
+- Split `Logger` service into `Logger` and `SystemLogger` (#3098)
+- Moved "Logs" from "General" tab to new "Logs" tab in the Settings (#3103)
+
+## Added
+
+- Enable logs by severity in the Settings (#3095)
+- Add Logs page to the menu, to browse the logs (#3100)
+- Split Logs by sources and date (#3100)
+- Logs count badge in the Logs menu link (#3105)
+
+### Improvements
+
+- Print the severity (`SUCCESS`/`INFO`/`WARNING`/`ERROR`) in the logs (#3099)
+- Use the Logger also when using plugin as standalone (without WordPress) (#3104)
+
+### Fixed
+
+- Error type for "Custom post does not exist" was missing from union types (#3109)
+
 ## 12.2.2 - 13/05/2025
 
 (Empty - Version sync with PRO)